requirements.txt
streamlit==1.24.0
openai==0.28.0
SpeechRecognition==3.8.1
pyttsx3==2.90
python-dotenv==0.19.0
Pillow==9.5.0 --only-binary :all:
